Implementing bulk shelf move - 1
There a bit of a bug now. The BookShelfChanger component does not reflect the books list for bulk move.
It only reflect when the search API return an error which hides the bulk action button. The when i search again with the correct key word, then select the books i mark for bulk action, it reflects the list.